% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_algo.R
\name{boot_algo.multclust}
\alias{boot_algo.multclust}
\title{function that implements the fast bootstrap algorithm as described in Roodman et al (2009)}
\usage{
\method{boot_algo}{multclust}(preprocessed_object)
}
\arguments{
\item{preprocessed_object}{A preprocessed object of time preprocessed_boottest}
}
\value{
A list of ...
}
\description{
function that implements the fast bootstrap algorithm as described in Roodman et al (2009)
}
