% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_algo.R
\name{boot_algo}
\alias{boot_algo}
\title{function that implements the bootstrap algorithm}
\usage{
boot_algo(preprocessed_object)
}
\arguments{
\item{preprocessed_object}{A preprocessed object of time preprocessed_boottest}
}
\value{
A list of ...
}
\description{
function that implements the bootstrap algorithm
}
