% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_algo.R
\name{boot_algo2.multclust}
\alias{boot_algo2.multclust}
\title{function that implements the fast bootstrap algorithm as described in Roodman et al (2009)}
\usage{
\method{boot_algo2}{multclust}(preprocessed_object, boot_iter, wild_draw_fun, ...)
}
\arguments{
\item{preprocessed_object}{A preprocessed object of time preprocessed_boottest}

\item{boot_iter}{number of bootstrap iterations}

\item{wild_draw_fun}{function. Specifies the type of bootstrap to use.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of ...
}
\description{
function that implements the fast bootstrap algorithm as described in Roodman et al (2009)
}
