% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_algo.R
\name{boot_algo2.oneclust}
\alias{boot_algo2.oneclust}
\title{function that implements the fast bootstrap algorithm as described in Roodman et al (2009)}
\usage{
\method{boot_algo2}{oneclust}(preprocessed_object, boot_iter)
}
\arguments{
\item{preprocessed_object}{A preprocessed object of time preprocessed_boottest}

\item{boot_iter}{number of bootstrap iterations}
}
\value{
A list of ...
}
\description{
function that implements the fast bootstrap algorithm as described in Roodman et al (2009)
}
