% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_algo2.R
\name{boot_algo2}
\alias{boot_algo2}
\title{function that implements the fast bootstrap algorithm as described in Roodman et al (2009)}
\usage{
boot_algo2(
  preprocessed_object,
  boot_iter,
  wild_draw_fun,
  point_estimate,
  impose_null,
  beta0,
  alpha,
  param,
  seed
)
}
\arguments{
\item{preprocessed_object}{A list: output of the preprocess2 function.}

\item{boot_iter}{number of bootstrap iterations}

\item{wild_draw_fun}{function. Specifies the type of bootstrap to use.}

\item{point_estimate}{The point estimate of the test parameter from the regression model.}

\item{impose_null}{If TRUE, the null is not imposed on the bootstrap distribution.
This is what Roodmal et al call the "WCU" bootstrap. With impose_null = FALSE, the
null is imposed ("WCR").}

\item{beta0}{Shifts the null hypothesis.}

\item{alpha}{The significance level.}

\item{param}{name of the test parameter.}

\item{seed}{the random seed. controls draw of bootstrap weights.}
}
\value{
A list of ...
}
\description{
function that implements the fast bootstrap algorithm as described in Roodman et al (2009)
}
