% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_algo3.R
\name{boot_algo3}
\alias{boot_algo3}
\title{function that implements the fast bootstrap algorithm as described in Roodman et al (2009)}
\usage{
boot_algo3(preprocessed_object, boot_iter, wild_draw_fun)
}
\arguments{
\item{preprocessed_object}{A preprocessed object of type preprocessed}

\item{boot_iter}{number of bootstrap iterations}

\item{wild_draw_fun}{function. Specifies the type of bootstrap to use.}
}
\value{
A list of ...
}
\description{
function that implements the fast bootstrap algorithm as described in Roodman et al (2009)
}
