% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_estimation.R
\name{boot_estimation}
\alias{boot_estimation}
\title{Bootstrap estimation function.
Estimates bootstrap p-values for all regression objects.}
\usage{
boot_estimation(X, Y, R0, clustid, B)
}
\arguments{
\item{X}{The matrix of covariates}

\item{Y}{The vector of outcome variables}

\item{R0}{The vector of constraints}

\item{B}{The number of bootstrap iterations}
}
\description{
Bootstrap estimation function.
Estimates bootstrap p-values for all regression objects.
}
