% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ssc.R
\name{boot_ssc}
\alias{boot_ssc}
\title{boot_ssc() handles the small sample correction factor applied in boottest() in
analogy to fixest::ssc()}
\usage{
boot_ssc(
  adj = FALSE,
  fixef.K = "none",
  cluster.adj = TRUE,
  cluster.df = "conventional"
)
}
\arguments{
\item{adj.}{Logical scalar, defaults}
}
\description{
boot_ssc() handles the small sample correction factor applied in boottest() in
analogy to fixest::ssc()
}
