% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_felm.R
\name{boottest.felm}
\alias{boottest.felm}
\title{Function that runs boottest for object of class felm}
\usage{
\method{boottest}{felm}(
  object,
  param,
  B,
  clustid,
  fe = NULL,
  conf_int = NULL,
  seed = NULL,
  beta0 = 0,
  alpha = NULL,
  type = "rademacher",
  ...
)
}
\arguments{
\item{object}{An object of class fixest}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{B}{number of bootstrap iterations}

\item{clustid}{A vector with the clusters}

\item{fe}{A character scalar. Fixed effect to be projected out in the bootstrap}

\item{conf_int}{A logical vector. If TRUE, boottest computes confidence intervals by p-value inversion}

\item{seed}{An integer. Allows the user to set a random seed}

\item{beta0}{A numeric. Shifts the null hypothesis}

\item{alpha}{A numeric between 0 and 1. E.g. alpha = 0.05 returns 0.95\% confidence intervals. By default, alpha = 0.05.}

\item{type}{character or function. The character string specifies the type of boostrap to use: One of "rademacher", "mammen", "norm" and "webb". Alternatively, type can be a function(n) for drawing wild bootstrap factors. "rademacher" by default.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class boottest
}
\description{
Function that runs boottest for object of class felm
}
