% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwc_boottest.R.R
\name{boottest}
\alias{boottest}
\title{Fast Wild Cluster Bootstrap}
\usage{
boottest(object, clustid, param, B, data, debug = FALSE, seed = NULL)
}
\arguments{
\item{object}{An regression object of type lm or lm_robust.}

\item{clustid}{A vector of clusters.}

\item{param}{The coefficient to be tested.}

\item{B}{The number of Bootstrap iterations.}

\item{debug}{If TRUE, sets debugging on.}

\item{seed}{Sets a specified seed.}
}
\description{
This function conducts a fast wild cluster bootstrap as in Cameron et al...
}
