% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_boottest_args_plus}
\alias{check_boottest_args_plus}
\title{several checks on the input args for 'boottest()'}
\usage{
check_boottest_args_plus(object, R, param, sign_level, B, fe = NULL)
}
\arguments{
\item{object}{an object of type lm, fixest or felm}

\item{R}{Numeric vector defining the test}

\item{param}{character vector - the parameters to be tested}

\item{sign_level}{The significance level}

\item{B}{the number of bootstrap iterations}

\item{fe}{NULL or numeric scalar - the fixed effect to be projected
out in the bootstrap}
}
\description{
several checks on the input args for 'boottest()'
}
