% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_mboottest_args_plus}
\alias{check_mboottest_args_plus}
\title{input argument tests for 'mboottest()'}
\usage{
check_mboottest_args_plus(object, R, r, fe)
}
\arguments{
\item{object}{an object of type lm, fixest or felm}

\item{R}{Numeric vector defining the test}

\item{r}{Numeric scalar, shifting the test}

\item{fe}{NULL or numeric scalar - the fixed effect to be projected
out in the bootstrap}
}
\description{
input argument tests for 'mboottest()'
}
