% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get_nthreads.R
\name{check_set_nthreads}
\alias{check_set_nthreads}
\title{Simple function that checks that the nber of threads is valid}
\usage{
check_set_nthreads(nthreads)
}
\arguments{
\item{nthreads}{Integer. Number of threads to be used}
}
\value{
Integer. The number of threads to be used.
}
\description{
Simple function that checks that the nber of threads is valid
}
