% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_get_nb_threads}
\alias{cpp_get_nb_threads}
\title{Get maximum number of threads on hardware for open mp support}
\usage{
cpp_get_nb_threads()
}
\value{
The maximum number of threads supported.
}
\description{
Get maximum number of threads on hardware for open mp support
}
