% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Simulate Data}
\usage{
create_data(N, N_G1, icc1, N_G2, icc2, numb_fe1, numb_fe2, seed, weights)
}
\arguments{
\item{N}{number of observations}

\item{N_G1}{A scalar. number of clusters for clustering variable 1}

\item{icc1}{A scalar between 0 and 1. intra-cluster correlation for clustering variable 1}

\item{N_G2}{A scalar. number of clusters for clustering variable 2}

\item{icc2}{A scalar between 0 and 1. intra-cluster correlation for clustering variable 2}

\item{numb_fe1}{A scalar. Number of fixed effect for first factor variable}

\item{numb_fe2}{A scalar. Number of fixed effect for second factor variable}

\item{seed}{An integer. Set the random seed}

\item{weights}{Possible regression weights to be used in estimation}
}
\value{
A simulated \code{data.frame} with specified numbers of clusters,
intra-cluster correlations and dimensionality of fixed effects.
}
\description{
Function simulates data for tests and examples with clustering variables
and fixed-effects.
}
