% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data_1}
\alias{create_data_1}
\title{Function creates data for tests and examples}
\usage{
create_data_1(N, N_G, icc, numb_fe1, numb_fe2)
}
\arguments{
\item{N}{number of observations}

\item{N_G}{number of clusters}

\item{icc}{intra-cluster correlation}

\item{numb_fe1}{A scalar. Number of fixed effect for first factor variable}

\item{numb_fe2}{A scalar. Number of fixed effect for second factor variable}
}
\description{
Function creates data for tests and examples
}
