% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data_2}
\alias{create_data_2}
\title{Function creates data for tests and examples}
\usage{
create_data_2(N, N_G1, icc1, N_G2, icc2, numb_fe1, numb_fe2)
}
\arguments{
\item{N}{number of observations}

\item{N_G}{number of clusters}

\item{icc}{intra-cluster correlation}
}
\description{
Function creates data for tests and examples
}
