% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab3}
\alias{crosstab3}
\title{collapse way to calculate crosstabs}
\usage{
crosstab3(data, var1, var2)
}
\arguments{
\item{data}{A matrix to collapse by two dimensions var1 var2}

\item{var1}{a data.frame containing a single variable}

\item{var2}{a data.frame containing a single variable}
}
\value{
A collapsed matrix of dimension length(unique(var1)) x length(unique(var2)). If...
}
\description{
collapse way to calculate crosstabs
}
