% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{setting options for nthreads when package is loaded}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\value{
Changes number of threads used.
}
\description{
setting options for nthreads when package is loaded
}
