% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Matrix Multiplication via Eigen}
\usage{
eigenMapMatMult(A, B, nthreads)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}
}
\value{
A matrix
}
\description{
Matrix Multiplication via Eigen
}
