% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMatMult}
\alias{eigenMatMult}
\title{Matrix Multiplication via Eigen}
\usage{
eigenMatMult(A, B)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}
}
\value{
A matrix
}
\description{
Matrix Multiplication via Eigen
}
