% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_proglang}
\alias{find_proglang}
\title{Check if julia or python are installed /
can be found on the PATH}
\usage{
find_proglang(lang)
}
\arguments{
\item{lang}{which language to check. Either 'julia' or 'python'}
}
\value{
logical. TRUE if lang is found on path, FALSE if not
}
\description{
Check if julia or python are installed /
can be found on the PATH
}
