% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwildclusterboot_setup.R
\name{fwildclusterboot_setup}
\alias{fwildclusterboot_setup}
\title{install Julia, connect R and Julia, install and pre-compile WildBootTests.jl}
\usage{
fwildclusterboot_setup()
}
\description{
install Julia, connect R and Julia, install and pre-compile WildBootTests.jl
}
