% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bootstrap_pvalue.R
\name{get_bootstrap_pvalue}
\alias{get_bootstrap_pvalue}
\title{get bootstrapped p-value based on bootstrapped t-stats}
\usage{
get_bootstrap_pvalue(p_val_type, t_stat, t_boot)
}
\arguments{
\item{p_val_type}{Character vector of length 1. Type of p-value.
Options include "two-tailed", "equal-tailed", ">" and "<".}

\item{t_stat}{The original t-statistic}

\item{t_boot}{The bootstrapped t-statistics}
}
\value{
A bootstrapped p-value
}
\description{
get bootstrapped p-value based on bootstrapped t-stats
}
