% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_seed.R
\name{get_seed}
\alias{get_seed}
\title{creates an integer based on the global random seed set via set.seed()
for using set.seed() for controlling rcpp's seed, see this
blog post http://rorynolan.rbind.io/2018/09/30/rcsetseed/}
\usage{
get_seed()
}
\description{
creates an integer based on the global random seed set via set.seed()
for using set.seed() for controlling rcpp's seed, see this
blog post http://rorynolan.rbind.io/2018/09/30/rcsetseed/
}
