% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ssc.R
\name{get_ssc}
\alias{get_ssc}
\title{Compute small sample adjustment factors}
\usage{
get_ssc(boot_ssc_object, N, k, G, vcov_sign)
}
\arguments{
\item{boot_ssc_object}{An object of type 'boot_ssc.type'}

\item{N}{The number of observations}

\item{k}{The number of estimated parameters}

\item{G}{The number of clusters}

\item{vcov_sign}{A vector that helps create the covariance matrix}
}
\value{
A small sample adjustment factor
}
\description{
Compute small sample adjustment factors
}
