% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{draw bootstrap weights}
\usage{
get_weights(type, full_enumeration, N_G_bootcluster, boot_iter)
}
\arguments{
\item{type}{the type of the weights distribution. Either 'rademacher',
'mammen', 'norm' or 'webb}

\item{full_enumeration}{Logical. Should (deterministic) full enumeration be
employed}

\item{N_G_bootcluster}{Integer. The number of bootstrap clusters}

\item{boot_iter}{The number of bootstrap iterations}
}
\value{
A matrix of dimension N_G_bootcluster x (boot_iter + 1)
}
\description{
draw bootstrap weights
}
