% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_p_val_fwc.R
\name{invert_p_val.algo_oneclust}
\alias{invert_p_val.algo_oneclust}
\title{Inverts the bootstrap p-value and calculates confidence sets}
\usage{
\method{invert_p_val}{algo_oneclust}(
  object,
  point_estimate,
  se_guess,
  clustid,
  fixed_effect,
  X,
  Y,
  N,
  k,
  param,
  R0,
  B,
  v,
  Xr,
  alpha,
  beta0,
  W,
  n_fe,
  N_G
)
}
\arguments{
\item{object}{A regression object of class lm, feols or felm}

\item{clustid}{A vector with the clusters}

\item{fixed_effect}{Fixed effect to be projected out in bootstrap}

\item{X}{the design matrix with the (potentially demeand) covariates}

\item{Y}{A numeric vector containing the outcome variable}

\item{N}{An integer. Number of observations}

\item{k}{An integer. Number of covariates (excluding fixed effects that are projected out)}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{R0}{A vector with the test constraint. Dimension (numb_covariates - numb_fe) x 1. 0 for covariate "param", else 1}

\item{B}{An integer. Number of bootstrap iterations}

\item{v}{A matrix. Draw from bootstrap distribution}

\item{Xr}{A matrix. R0 \%*\% X}

\item{alpha}{A numeric between 0 and 1. Sets to confidence level: alpha = 0.05 returns 0.95\% confidence intervals}

\item{seed}{An integer. Seed.}

\item{N_g}{An integer. Number of clusters.}

\item{XinvXXr}{A matrix. see boottest()}

\item{SXinvXXRX}{A matrix. see boottest() for computation}
}
\description{
Inverts the bootstrap p-value and calculates confidence sets
}
