% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_p_val_fwc.R
\name{invert_p_val2.algo_multclust}
\alias{invert_p_val2.algo_multclust}
\title{Inverts the bootstrap p-value and calculates confidence sets}
\usage{
\method{invert_p_val2}{algo_multclust}(object, B, point_estimate, se_guess, clustid, alpha)
}
\arguments{
\item{object}{A regression object of class lm, feols or felm}
}
\description{
Inverts the bootstrap p-value and calculates confidence sets
}
