% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{invert_p_val2}
\alias{invert_p_val2}
\title{S3 Method invert_p_val2}
\usage{
invert_p_val2(x, ...)
}
\arguments{
\item{x}{An object of type boot_algo_oneclust or boot_algo_multclust}

\item{...}{other arguments}
}
\description{
S3 Method invert_p_val2
}
