% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{invert_p_val2a}
\alias{invert_p_val2a}
\title{S3 Method invert_p_val2}
\usage{
invert_p_val2a(object, ...)
}
\arguments{
\item{object}{An object of type boot_algo_oneclust or boot_algo_multclust}

\item{...}{other arguments}
}
\description{
S3 Method invert_p_val2
}
