% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_p_val_fwc.R
\name{invert_p_val_fwc}
\alias{invert_p_val_fwc}
\title{Inverts the bootstrap p-value and calculates confidence sets}
\usage{
invert_p_val_fwc(
  object,
  data,
  clustid,
  X,
  Y,
  param,
  R0,
  B,
  N,
  k,
  seed,
  N_g,
  invXX,
  v,
  Xr,
  XinvXXr,
  SXinvXXRX,
  alpha
)
}
\arguments{
\item{object}{A regression object of class lm, feols or felm}

\item{data}{A data.frame with the data used in the estimation. If fixed effects are used, this data is demeaned and excludes the fixed effects}

\item{clustid}{A vector with the clusters}

\item{X}{the design matrix with the (potentially demeand) covariates}

\item{Y}{A numeric vector containing the outcome variable}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{R0}{A vector with the test constraint. Dimension (numb_covariates - numb_fe) x 1. 0 for covariate "param", else 1}

\item{B}{An integer. Number of bootstrap iterations}

\item{N}{An integer. Number of observations}

\item{k}{An integer. Number of covariates (excluding fixed effects that are projected out)}

\item{seed}{An integer. Seed.}

\item{N_g}{An integer. Number of clusters.}

\item{v}{A matrix. Draw from bootstrap distribution}

\item{Xr}{A matrix. R0 \%*\% X}

\item{XinvXXr}{A matrix. see boottest()}

\item{SXinvXXRX}{A matrix. see boottest() for computation}

\item{alpha}{A numeric between 0 and 1. Sets to confidence level: alpha = 0.05 returns 0.95\% confidence intervals}
}
\description{
Inverts the bootstrap p-value and calculates confidence sets
}
