% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model_matrix.felm}
\alias{model_matrix.felm}
\title{Enhanced model.matrix for objects of type felm}
\usage{
\method{model_matrix}{felm}(object, type, collin.rm = TRUE, ...)
}
\arguments{
\item{object}{An object of class felm}

\item{type}{'rhs' for right-hand side variables, 'fixef' for fixed effects}

\item{collin.rm}{Should collinear variables be dropped?}

\item{...}{Other arguments}
}
\description{
Enhanced model.matrix for objects of type felm
}
