% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model_matrix.fixest}
\alias{model_matrix.fixest}
\title{Enhanced model.matrix for objects of type fixest}
\usage{
\method{model_matrix}{fixest}(object, type, collin.rm = TRUE, ...)
}
\arguments{
\item{object}{An object of class fixest}

\item{type}{rhs lhs or fixef}

\item{collin.rm}{Should collinear variables be dropped?}

\item{...}{Other arguments}
}
\description{
Enhanced model.matrix for objects of type fixest
}
