% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix2.R
\name{model_matrix.ivreg}
\alias{model_matrix.ivreg}
\title{Enhanced model.matrix for objects of type ivreg}
\usage{
\method{model_matrix}{ivreg}(object, collin.rm = TRUE, type)
}
\description{
Enhanced model.matrix for objects of type ivreg
}
