% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model_matrix.lm}
\alias{model_matrix.lm}
\title{Enhanced model.matrix for objects of type lm}
\usage{
\method{model_matrix}{lm}(object, collin.rm = TRUE, ...)
}
\arguments{
\item{object}{An object of class lm}

\item{collin.rm}{Should collinear variables be dropped?}

\item{...}{Other arguments}
}
\description{
Enhanced model.matrix for objects of type lm
}
