% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model_matrix.plm}
\alias{model_matrix.plm}
\title{Enhanced model.matrix for objects of type plm}
\usage{
\method{model_matrix}{plm}(object, collin.rm = TRUE, ...)
}
\arguments{
\item{object}{An object of class plm}

\item{collin.rm}{Should collinear variables be dropped?}

\item{...}{Other arguments}
}
\description{
Enhanced model.matrix for objects of type plm
}
