% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model_matrix}
\alias{model_matrix}
\title{enhanced model.matrix functionalities}
\usage{
model_matrix(object, ...)
}
\arguments{
\item{object}{An object of class \code{lm} or `felm``}

\item{...}{Other arguments}
}
\description{
enhanced model.matrix functionalities
}
