% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_val_null.R
\name{p_val_null2}
\alias{p_val_null2}
\title{Calculate p-values based on A, B, CC, CD, DD and other inputs}
\usage{
p_val_null2(
  beta0,
  A,
  B,
  CC,
  CD,
  DD,
  clustid,
  boot_iter,
  small_sample_correction
)
}
\arguments{
\item{beta0}{Scalar. Shifts the null hypothesis.}

\item{A}{A list.}

\item{B}{A list.}

\item{CC}{A list.}

\item{CD}{A list.}

\item{DD}{A list.}

\item{clustid}{A data.frame containing the cluster variables.}

\item{boot_iter}{An integer. Number of bootstrap iterations.}

\item{small_sample_correction}{A vector of the dimension of ncol(clustid)}
}
\description{
Calculate p-values based on A, B, CC, CD, DD and other inputs
}
