% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{plot}
\alias{plot}
\title{S3 generic to plot bootstrap t statistics}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{object of type boottest}

\item{...}{other arguments}
}
\description{
S3 generic to plot bootstrap t statistics
}
