% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{plot_boot}
\alias{plot_boot}
\title{Plot the bootstrap distribution of t-statistics}
\usage{
plot_boot(object)
}
\arguments{
\item{object}{An object of type boottest}
}
\description{
Plot the bootstrap distribution of t-statistics
}
