% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_felm.R
\name{preprocess.felm}
\alias{preprocess.felm}
\title{function that pre-processes regression objects of type fixest}
\usage{
\method{preprocess}{felm}(object, param, clustid, beta0, alpha, fe, seed)
}
\arguments{
\item{object}{An object of class fixest}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{clustid}{A vector with the clusters}

\item{beta0}{A numeric. Shifts the null hypothesis}

\item{alpha}{A numeric between 0 and 1. Sets to confidence level: alpha = 0.05 returns 0.95\% confidence intervals}
}
\value{
preprocessed object of class boottest_preprocessed
}
\description{
function that pre-processes regression objects of type fixest
}
