% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess.lm}
\alias{preprocess.lm}
\title{function that pre-processes regression objects of type lm}
\usage{
\method{preprocess}{lm}(object, param, clustid, beta0, alpha, seed)
}
\arguments{
\item{object}{An object of class lm}

\item{param}{A scalar character. The coefficient for which a hypothesis is to be tested}

\item{clustid}{A character vector containing the names of the cluster variables}

\item{beta0}{A numeric scalar. Shifts the null hypothesis}

\item{alpha}{A numeric scalar between 0 and 1. Sets to confidence level: alpha = 0.05 returns 0.95\% confidence intervals}

\item{seed}{An integer. Sets the seed}
}
\value{
preprocessed object of class boottest_preprocessed
}
\description{
function that pre-processes regression objects of type lm
}
