% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{preprocess}
\alias{preprocess}
\title{S3 Method preprocess}
\usage{
preprocess(object, ...)
}
\arguments{
\item{object}{An object of type lm, fixest of felm}

\item{...}{other arguments}
}
\description{
S3 Method preprocess
}
