% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_lm.R
\name{preprocess_lm}
\alias{preprocess_lm}
\title{function that pre-processes regression objects of type lm}
\usage{
preprocess_lm(object, param, clustid, beta0, alpha)
}
\arguments{
\item{object}{An object of class lm}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{clustid}{A vector with the clusters}

\item{beta0}{A numeric. Shifts the null hypothesis}

\item{alpha}{A numeric between 0 and 1. Sets to confidence level: alpha = 0.05 returns 0.95\% confidence intervals}

\item{demean}{If TRUE, fixed effects are projected out prior to the bootstrap. FALSE by default}
}
\value{
preprocessed object of class boottest_preprocessed
}
\description{
function that pre-processes regression objects of type lm
}
