% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mboottest}
\alias{print.mboottest}
\title{S3 method to print key information for objects of type \code{mboottest}}
\usage{
\method{print}{mboottest}(x, ..., digits = 4)
}
\arguments{
\item{x}{object of type mboottest}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A scalar containing the effective number of observations used in \code{mboottest}
}
\description{
S3 method to print key information for objects of type \code{mboottest}
}
