% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pval}
\alias{pval}
\title{\code{pval} is a S3 method to collect pvalues for objects
of type \code{boottest} and \code{mboottest}}
\usage{
pval(object, ...)
}
\arguments{
\item{object}{An object of type lm, fixest, felm or ivreg}

\item{...}{other arguments}
}
\description{
\code{pval} is a S3 method to collect pvalues for objects
of type \code{boottest} and \code{mboottest}
}
