% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secant_rel.R
\name{secant_rel}
\alias{secant_rel}
\title{secant method
function adopted from NLRoot::SMfzero}
\usage{
secant_rel(f, x1, x2, B, num = 10, eps = 1e-06)
}
\description{
secant method
function adopted from NLRoot::SMfzero
}
