% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_boottest_seed.R
\name{set_boottest_seed}
\alias{set_boottest_seed}
\title{Set a seed for fwildclusterboot}
\usage{
set_boottest_seed(seed, set_julia_seed = FALSE)
}
\arguments{
\item{seed}{Integer. Sets a random seed}

\item{WildBootTests.jl}{Logical. Should a Julia seed be set? Requires
installation of Julia and JuliaConnectoR. FALSE by default.}
}
\description{
Set a seed for fwildclusterboot
}
