% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_julia_seed.R
\name{set_julia_seed}
\alias{set_julia_seed}
\title{Set a global seed in the current Julia session}
\usage{
set_julia_seed(rng)
}
\arguments{
\item{rng}{An integer that controls the random number generation}
}
\description{
Set a global seed in the current Julia session
}
