% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get_Boottest_boot_algo.R
\name{setBoottest_boot_algo}
\alias{setBoottest_boot_algo}
\title{Sets the bootstrap algo to be run via \code{boottest()} and \code{waldboottest()}}
\usage{
setBoottest_boot_algo(boot_algo)
}
\arguments{
\item{boot_algo}{Character scalar. Either 'R' or 'WildBootTests.jl'. Default is 'R'}
}
\value{
No return value
}
\description{
Sets the bootstrap algo to be run via \code{boottest()} and \code{waldboottest()}
}
\examples{
\dontrun{
setBoottest_boot_algo(boot_algo = "R")
setBoottest_boot_algo(boot_algo = "WildBootTests.jl")
}
}
