% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get_internal_seed.R
\name{setBoottest_internal_seed}
\alias{setBoottest_internal_seed}
\title{Set global variable whether \code{boottest()} / \code{mboottest()} should use an internal seed.}
\usage{
setBoottest_internal_seed(internal_seed)
}
\arguments{
\item{internal_seed}{Logical scalar. If TRUE, use internal seed for all
calls of boottest() unless specified otherwise}
}
\description{
Set global variable whether \code{boottest()} / \code{mboottest()} should use an internal seed.
}
