% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get_nthreads.R
\name{setBoottest_nthreads}
\alias{setBoottest_nthreads}
\title{Set the number of threads for use with open m
By default, we use only 50\% of threads (never use all)}
\usage{
setBoottest_nthreads(nthreads)
}
\arguments{
\item{nthreads}{Integer. Number of threads to be used}
}
\description{
Set the number of threads for use with open m
By default, we use only 50\% of threads (never use all)
}
