% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{summarize_boot.boottest}
\alias{summarize_boot.boottest}
\title{S3 method to summarize objects of class boottest}
\usage{
\method{summarize_boot}{boottest}(object, digits = 3)
}
\arguments{
\item{object}{object of type boottest}

\item{digits}{rounding of output}
}
\description{
S3 method to summarize objects of class boottest
}
