% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{summarize_boot}
\alias{summarize_boot}
\title{S3 method to summarize objects of class boottest}
\usage{
summarize_boot(x, ...)
}
\arguments{
\item{x}{object of type boottest}

\item{...}{other arguments}
}
\description{
S3 method to summarize objects of class boottest
}
