% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{summary}
\alias{summary}
\title{S3 method to summarize objects of class boottest}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{object of type boottest}

\item{...}{other arguments}
}
\description{
S3 method to summarize objects of class boottest
}
