% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldtest_methods.R
\name{summary.waldboottest}
\alias{summary.waldboottest}
\title{S3 method to summarize objects of class waldboottest}
\usage{
\method{summary}{waldboottest}(object, digits = 3, ...)
}
\arguments{
\item{object}{object of type waldboottest}

\item{digits}{rounding of output. 3 by default}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Returns result summaries for objects of type waldboottest
}
\description{
S3 method to summarize objects of class waldboottest
}
