% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_fixest.R
\name{sunab}
\alias{sunab}
\title{this is a trick - actually not need
it's just that something called sunab() needs to be
found, else cluster creation funs confused}
\usage{
sunab(
  cohort,
  period,
  ref.c = NULL,
  ref.p = -1,
  bin,
  bin.rel,
  bin.c,
  bin.p,
  att = FALSE,
  no_agg = FALSE
)
}
\description{
this is a trick - actually not need
it's just that something called sunab() needs to be
found, else cluster creation funs confused
}
