% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{teststat.boottest}
\alias{teststat.boottest}
\title{S3 method to obtain the non-bootstrapped t-statistic calculated via \code{boottest()}}
\usage{
\method{teststat}{boottest}(object, ...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}
}
\value{
A vector containing the non-bootstrapped t-statistic calculated in \code{boottest()}
}
\description{
S3 method to obtain the non-bootstrapped t-statistic calculated via \code{boottest()}
}
\examples{
library(fwildclusterboot)
data(voters)
data = voters
)
boot1 <- boottest(lm_fit,
  B = 9999,
 param = "treatment",
   clustid = "group_id1"
 )
teststat(boot1)
}
