% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{teststat}
\alias{teststat}
\title{\code{teststat} is a S3 method to collect teststats for objects
of type \code{boottest} and \code{mboottest}}
\usage{
teststat(object, ...)
}
\arguments{
\item{object}{An object of type lm, fixest, felm or ivreg}

\item{...}{other arguments}
}
\description{
\code{teststat} is a S3 method to collect teststats for objects
of type \code{boottest} and \code{mboottest}
}
