% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{tidy.boottest}
\alias{tidy.boottest}
\title{tidy <- function(x, ...){
#'@export
UseMethod("tidy")
}}
\usage{
\method{tidy}{boottest}(object)
}
\description{
summary <- function(x, ...){
#'@export
UseMethod("summary")
}
}
