% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldtest_methods.R
\name{tidy.mboottest}
\alias{tidy.mboottest}
\title{S3 method to summarize objects of class mboottest into tidy data.frame}
\usage{
\method{tidy}{mboottest}(object, ...)
}
\arguments{
\item{object}{object of type mboottest}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A tidy data.frame with estimation results for objects of type
mboottest
}
\description{
S3 method to summarize objects of class mboottest into tidy data.frame
}
