% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy}
\alias{tidy}
\title{S3 method to summarize objects of class boottest into tidy data.frame}
\usage{
tidy(object, ...)
}
\arguments{
\item{object}{object of type boottest}

\item{...}{other arguments}
}
\description{
S3 method to summarize objects of class boottest into tidy data.frame
}
