% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldtest_methods.R
\name{tidy.waldboottest}
\alias{tidy.waldboottest}
\title{S3 method to summarize objects of class waldboottest into tidy data.frame}
\usage{
\method{tidy}{waldboottest}(object, ...)
}
\arguments{
\item{object}{object of type waldboottest}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A tidy data.frame with estimation results for objects of type
waldboottest
}
\description{
S3 method to summarize objects of class waldboottest into tidy data.frame
}
