% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_integer.R
\name{to_integer}
\alias{to_integer}
\title{Transform vectors of all types safely to integer vectors}
\usage{
to_integer(vec)
}
\arguments{
\item{vec}{A vector}
}
\value{
An integer vector
}
\description{
Transform vectors of all types safely to integer vectors
}
