% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{transform_fe2}
\alias{transform_fe2}
\title{preprocess the model fixed effects. If is.null(fe) == TRUE, all
fixed effects specified in the fixest or felm model are simply added
as dummy variables to the design matrix X. If !is.null(fe), the fixed
effect specified via fe is projected out in the bootstrap - all other
fixed effects are added as dummy variables}
\usage{
transform_fe2(object, X, Y, fe, has_weights, N, boot_algo)
}
\description{
preprocess the model fixed effects. If is.null(fe) == TRUE, all
fixed effects specified in the fixest or felm model are simply added
as dummy variables to the design matrix X. If !is.null(fe), the fixed
effect specified via fe is projected out in the bootstrap - all other
fixed effects are added as dummy variables
}
