% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voters-data.R
\docType{data}
\name{voters}
\alias{voters}
\title{Random example data set created via create_data_2 function}
\format{
An object of class \code{data.frame} with 300 rows and 13 columns.
}
\usage{
data(voters)
}
\description{
Random example data set created via create_data_2 function
}
\examples{
data(voters)
}
\keyword{datasets}
