% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boot_algo}
\alias{boot_algo}
\title{S3 Method boot_algo}
\usage{
boot_algo(x, ...)
}
\arguments{
\item{x}{An object of type preprocess_lm, preprocess_felm or preprocess_fixest}

\item{...}{other arguments}
}
\description{
S3 Method boot_algo
}
