% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boot_algo2}
\alias{boot_algo2}
\title{S3 Method boot_algo2}
\usage{
boot_algo2(x, ...)
}
\arguments{
\item{x}{An object of type preprocess_lm, preprocess_felm or preprocess_fixest}

\item{...}{other arguments}
}
\description{
S3 Method boot_algo2
}
