% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootagg.R
\name{bootagg}
\alias{bootagg}
\title{Aggregates the value of coefficients. Inference via a wild cluster
bootstrap
Most of this function is written by Laurent Bergé and used
in the fixest package published under GPL-3,
https://cran.r-project.org/web/packages/fixest/index.html
minor changes by Alexander Fischer}
\usage{
bootagg(
  x,
  agg,
  full = FALSE,
  use_weights = TRUE,
  B = 999,
  bootstrap_type = "11",
  clustid,
  nthreads = 1,
  ...
)
}
\description{
Aggregates the value of coefficients. Inference via a wild cluster
bootstrap
Most of this function is written by Laurent Bergé and used
in the fixest package published under GPL-3,
https://cran.r-project.org/web/packages/fixest/index.html
minor changes by Alexander Fischer
}
\examples{
library(fixest)
library(fwildclusterboot)

set.seed(123456L)

# code by Grant McDermott
# 60 time periods, 30 individuals, and 5 waves of treatment
tmax = 60; imax = 30; nlvls = 5

dat = 
  expand.grid(time = 1:tmax, id = 1:imax) |>
  within({
    
    cohort      = NA
    effect      = NA
    first_treat = NA
    
    for (chrt in 1:imax) {
      cohort = ifelse(id==chrt, sample.int(nlvls, 1), cohort)
    }
    
    for (lvls in 1:nlvls) {
      effect      = ifelse(cohort==lvls, sample(2:10, 1), effect)
      first_treat = ifelse(cohort==lvls, sample(1:(tmax+20), 1), first_treat)
    }
    
    first_treat = ifelse(first_treat>tmax, Inf, first_treat)
    treat       = time>=first_treat
    rel_time    = time - first_treat
    y           = id + time + ifelse(treat, effect*rel_time, 0) + rnorm(imax*tmax)
    
    rm(chrt, lvls, cohort, effect)
  })

head(dat)
dim(dat)

object = feols(
  y ~ sunab(first_treat, rel_time) + id + time, 
  data = dat, vcov = ~id
)

x = object
agg = TRUE
full = FALSE
use_weights = TRUE
B = 999
clustid = ~time
bootstrap_type = "11"
nthreads = 2

res <- 
  bootagg(
    x,
    agg = TRUE,
    full = FALSE,
    use_weights = TRUE,
    B = 999, 
    bootstrap_type = "11", 
    clustid,
    nthreads = 8
  )

res
sunab_p <- pvalue(summary(x, agg = TRUE))
cbind(
  res, 
  round(sunab_p[-94][-93][-1], 3)
)
}
