% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_felm.R
\name{boottest.felm}
\alias{boottest.felm}
\title{Function that runs boottest for object of class felm}
\usage{
\method{boottest}{felm}(
  object,
  clustid,
  param,
  B,
  data,
  alpha = NULL,
  fixed_effects = NULL,
  weights = NULL,
  conf_int = NULL,
  debug = FALSE,
  seed = NULL,
  beta0 = 0
)
}
\arguments{
\item{object}{an object of type felm}

\item{clustid}{a vector of clusters}

\item{param}{the paramter that is to be tested}

\item{B}{number of bootstrap iterations}

\item{data}{data}

\item{fixed_effects}{optional argument}
}
\value{
An object of class boottest
}
\description{
Function that runs boottest for object of class felm
}
