% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_felm.R
\name{boottest.felm}
\alias{boottest.felm}
\title{Conducts wild cluster bootstrap inference for object of class felm.}
\usage{
\method{boottest}{felm}(
  object,
  param,
  B,
  clustid,
  bootcluster = "max",
  fe = NULL,
  conf_int = NULL,
  seed = NULL,
  beta0 = 0,
  alpha = NULL,
  type = "rademacher",
  ...
)
}
\arguments{
\item{object}{An object of class fixest}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{B}{number of bootstrap iterations}

\item{clustid}{A vector with the clusters}

\item{bootcluster}{A character vector. Sets the cluster used in the boottest. Chooses the largest cluster by default}

\item{fe}{A character scalar. Fixed effect to be projected out in the bootstrap}

\item{conf_int}{A logical vector. If TRUE, boottest computes confidence intervals by p-value inversion}

\item{seed}{An integer. Allows the user to set a random seed}

\item{beta0}{A numeric. Shifts the null hypothesis}

\item{alpha}{A numeric between 0 and 1. E.g. alpha = 0.05 returns 0.95\% confidence intervals. By default, alpha = 0.05.}

\item{type}{character or function. The character string specifies the type of boostrap to use: One of "rademacher", "mammen", "norm" and "webb". Alternatively, type can be a function(n) for drawing wild bootstrap factors. "rademacher" by default.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class \code{boottest}
\item{p_val}{The bootstrap p-value.}
\item{t_stat}{The bootstrap t-statistic.}
\item{conf_int}{The bootstrap confidence interval.}
\item{param}{The tested parameter.}
\item{N}{Sample size. Might differ from the regression sample size if the cluster variables contain NA values.}
\item{B}{Number of Bootstrap Iterations.}
\item{clustid}{Names of the cluster Variables.}
\item{N_G}{Dimension of the cluster variables as used in boottest.}
\item{alpha}{Significance level used in boottest.}
\item{type}{Distribution of the bootstrap weights.}
\item{p_test_vals}{All p-values calculated while calculating the confidence interval.}
\item{test_vals}{All t-statistics calculated while calculating the confidence interval.}
\item{regression}{The regression object used in boottest.}
\item{call}{Function call of boottest.}
}
\description{
Conducts wild cluster bootstrap inference for object of class felm.
}
\section{Confidence Intervals}{

\code{boottest} computes confidence intervals by inverting p-values. In practice, the following procedure is used:
\itemize{
\item Based on an initial guess for starting values, calculate p-values for 26 equal spaced points between the starting values.
\item Out of the 26 calculated p-values, find the two pairs of values x for which the corresponding p-values px cross the significance level alpha.
\item Feed the two pairs of x into an numerical root finding procedure and solve for the root. boottest currently relies on \code{stats::uniroot} and sets an absolute tolerance of 1e-06 and stops the procedure after 10 iterations.
}
Note that confidence intervals computed via \code{boottest.felm} sometimes slightly differ from confidence intervals calculated
via \code{boottest.lm} or \code{boottest.fixest}. This is due to different initial guesses for starting values. \code{boottest.lm} uses inflated cluster robust
standard errors calculated via the \code{sandwich} package, while \code{boottest.felm} and \code{boottest.fixest} use different small-sample bias
correction methods for calculating cluster standard errors internally. Slightly different starting values hence lead to slightly different
confidence intervals.
}

\section{Standard Errors}{

\code{boottest} does not calculate standard errors.
}

\references{
Roodman et al., 2019, "Fast and wild: Bootstrap inference in Stata using boottest", The Stata Journal. (\url{https://journals.sagepub.com/doi/full/10.1177/1536867X19830877})
}
