% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_felm.R
\name{boottest.felm}
\alias{boottest.felm}
\title{Function that runs boottest for object of class felm}
\usage{
\method{boottest}{felm}(
  object,
  clustid,
  param,
  B,
  fe = NULL,
  weights = NULL,
  conf_int = NULL,
  debug = FALSE,
  seed = NULL,
  beta0 = 0,
  alpha = NULL
)
}
\arguments{
\item{object}{An object of class fixest}

\item{clustid}{A vector with the clusters}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{B}{number of bootstrap iterations}

\item{weights}{Regression weights. Currently, WLS is not supported, and weights needs to be NULL}

\item{conf_int}{A logical vector. If TRUE, boottest computes confidence intervals by p-value inversion}

\item{seed}{An integer. Allows the user to set a random seed}

\item{beta0}{A numeric. Shifts the null hypothesis}

\item{alpha}{A numeric between 0 and 1. Sets to confidence level: alpha = 0.05 returns 0.95\% confidence intervals}
}
\value{
An object of class boottest
}
\description{
Function that runs boottest for object of class felm
}
