% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_fixest.R
\name{boottest.fixest}
\alias{boottest.fixest}
\title{Computes wild cluster bootstrap for object of class fixest}
\usage{
\method{boottest}{fixest}(
  object,
  clustid,
  param,
  B,
  fe = NULL,
  alpha = NULL,
  conf_int = NULL,
  seed = NULL,
  beta0 = 0,
  type = "rademacher",
  ...
)
}
\arguments{
\item{object}{An object of class fixest. Note: advanced formula tools in fixest or vectorized formulas are currently not supported.}

\item{clustid}{A vector with the clusters}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{B}{number of bootstrap iterations}

\item{fe}{A character scalar. Fixed effect to be projected out in the bootstrap}

\item{alpha}{A numeric between 0 and 1. E.g. alpha = 0.05 returns 0.95\% confidence intervals. By default, alpha = 0.05.}

\item{conf_int}{A logical vector. If TRUE, boottest computes confidence intervals by p-value inversion}

\item{seed}{An integer. Allows the user to set a random seed}

\item{beta0}{A numeric. Shifts the null hypothesis}

\item{type}{character or function. The character string specifies the type of boostrap to use: One of "rademacher", "mammen", "norm" and "webb". Alternatively, type can be a function(n) for drawing wild bootstrap factors. "rademacher" by default.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class boottest
}
\description{
Computes wild cluster bootstrap for object of class fixest
}
\examples{
library(fwildclusterboot)
library(fixest)
voters <- create_data_2(N = 10000, N_G1 = 20, icc1 = 0.91, N_G2 = 10, icc2 = 0.51, numb_fe1 = 10, numb_fe2 = 10, seed = 12345)
feols_fit <-feols(proposition_vote ~ treatment + ideology1 + log_income, fixef =  "Q1_immigration", weights = NULL, data = voters)
boot1 <- boottest(feols_fit, B = 10000, param = "treatment", clustid = "group_id1")
boot2 <- boottest(feols_fit, B = 10000, param = "treatment", clustid = c("group_id1", "group_id2"))
boot3 <- boottest(feols_fit, B = 10000, param = "treatment", clustid = c("group_id1", "group_id2"), fe = "Q1_immigration")
boot4 <- boottest(feols_fit, B = 10000, param = "treatment", clustid = c("group_id1", "group_id2"), fe = "Q1_immigration", alpha = 0.2, seed = 8, beta0 = 2)
summary(boot1)
tidy(boot1)
plot(boot1)
}
