% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_lm.R
\name{boottest.lm}
\alias{boottest.lm}
\title{Function that runs boottest for object of class lm}
\usage{
\method{boottest}{lm}(
  object,
  clustid,
  param,
  B,
  weights = NULL,
  conf_int = NULL,
  seed = NULL,
  beta0 = NULL,
  alpha = NULL,
  type = "rademacher",
  ...
)
}
\arguments{
\item{object}{An object of class lm}

\item{clustid}{A vector with the clusters}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{B}{number of bootstrap iterations}

\item{weights}{Regression weights. Currently, WLS is not supported, and weights needs to be NULL}

\item{conf_int}{A logical vector. If TRUE, boottest computes confidence intervals by p-value inversion}

\item{seed}{An integer. Allows the user to set a random seed}

\item{beta0}{A numeric. Shifts the null hypothesis}

\item{alpha}{A numeric between 0 and 1. Sets to confidence level: alpha = 0.05 returns 0.95\% confidence intervals}

\item{type}{character or function. The character string specifies the type of boostrap to use: One of "rademacher", "mammen", "norm" and "webb". Alternatively, type can be a function(n) for drawing wild bootstrap factors. "rademacher" by default.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class boottest
}
\description{
Function that runs boottest for object of class lm
}
\examples{
library(fwildclusterboot)
voters <- create_data_2(N = 10000, N_G1 = 20, icc1 = 0.91, N_G2 = 10, icc2 = 0.51, numb_fe1 = 10, numb_fe2 = 10, seed = 12345)
lm_fit <-lm(proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration, weights = NULL, data = voters)
boot1 <- boottest(lm_fit, B = 10000, param = "treatment", clustid = "group_id1")
boot2 <- boottest(lm_fit, B = 10000, param = "treatment", clustid = c("group_id1", "group_id2"))
boot3 <- boottest(lm_fit, B = 10000, param = "treatment", clustid = c("group_id1", "group_id2"), alpha = 0.2, seed = 8, beta0 = 2)
summary(boot1)
tidy(boot1)
plot(boot1)
}
