% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_lm.R
\name{boottest.lm}
\alias{boottest.lm}
\title{Conducts wild cluster bootstrap inference for object of class lm.}
\usage{
\method{boottest}{lm}(
  object,
  clustid,
  param,
  B,
  conf_int = NULL,
  seed = NULL,
  beta0 = NULL,
  alpha = NULL,
  type = "rademacher",
  ...
)
}
\arguments{
\item{object}{An object of class lm}

\item{clustid}{A vector with the clusters}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{B}{number of bootstrap iterations}

\item{conf_int}{A logical vector. If TRUE, boottest computes confidence intervals by p-value inversion}

\item{seed}{An integer. Allows the user to set a random seed}

\item{beta0}{A numeric. Shifts the null hypothesis}

\item{alpha}{A numeric between 0 and 1. E.g. alpha = 0.05 returns 0.95\% confidence intervals. By default, alpha = 0.05.}

\item{type}{character or function. The character string specifies the type of boostrap to use: One of "rademacher", "mammen", "norm" and "webb". Alternatively, type can be a function(n) for drawing wild bootstrap factors. "rademacher" by default.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class \code{boottest}
\item{p_val}{The bootstrap p-value.}
\item{t_stat}{The bootstrap t-statistic.}
\item{conf_int}{The bootstrap confidence interval.}
\item{param}{The tested parameter.}
\item{N}{Sample size. Might differ from the regression sample size if the cluster variables contain NA values.}
\item{B}{Number of Bootstrap Iterations.}
\item{clustid}{Names of the cluster Variables.}
\item{N_G}{Dimension of the cluster variables as used in boottest.}
\item{alpha}{Significance level used in boottest.}
\item{type}{Distribution of the bootstrap weights.}
\item{p_test_vals}{All p-values calculated while calculating the confidence interval.}
\item{test_vals}{All t-statistics calculated while calculating the confidence interval.}
\item{regression}{The regression object used in boottest.}
\item{call}{Function call of boottest.}
}
\description{
Conducts wild cluster bootstrap inference for object of class lm.
}
\section{Confidence Intervals}{

\code{boottest} computes confidence intervals by inverting p-values. In practice, the following procedure is used:
\itemize{
\item Based on an initial guess for starting values, calculate p-values for 26 equal spaced points between the starting values.
\item Out of the 26 calculated p-values, find the two pairs of values x for which the corresponding p-values px cross the significance level alpha.
\item Feed the two pairs of x into an numerical root finding procedure and solve for the root. boottest currently relies on \code{stats::uniroot} and sets an absolute tolerance of 1e-06 and stops the procedure after 10 iterations.
}
}

\section{Standard Errors}{

\code{boottest} does not calculate standard errors.
}

\examples{
library(fwildclusterboot)
voters <- create_data_2(N = 10000, N_G1 = 20, icc1 = 0.91, N_G2 = 10, icc2 = 0.51, numb_fe1 = 10, numb_fe2 = 10, seed = 12345)
lm_fit <-lm(proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration, data = voters)
boot1 <- boottest(lm_fit, B = 10000, param = "treatment", clustid = "group_id1")
boot2 <- boottest(lm_fit, B = 10000, param = "treatment", clustid = c("group_id1", "group_id2"))
boot3 <- boottest(lm_fit, B = 10000, param = "treatment", clustid = c("group_id1", "group_id2"), alpha = 0.2, seed = 8, beta0 = 2)
summary(boot1)
tidy(boot1)
plot(boot1)
}
\references{
Roodman et al., 2019, "Fast and wild: Bootstrap inference in Stata using boottest", The Stata Journal. (\url{https://journals.sagepub.com/doi/full/10.1177/1536867X19830877})
}
