% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_lm_robust.R, R/fwc_boottest.R.R
\name{boottest.lm_robust}
\alias{boottest.lm_robust}
\title{method for object of class "lm_robust"}
\usage{
boottest.lm_robust(object, clustid, param, B, data, debug = FALSE, seed = NULL)

boottest.lm_robust(object, clustid, param, B, data, debug = FALSE, seed = NULL)
}
\description{
method for object of class "lm_robust"

method for object of class "lm_robust"
}
