% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest_lm_robust.R
\name{boottest.lm_robust}
\alias{boottest.lm_robust}
\title{method for object of class "lm_robust"}
\usage{
boottest.lm_robust(
  object,
  param,
  B,
  data,
  clustid,
  weights = NULL,
  conf_int = NULL,
  debug = FALSE,
  seed = NULL,
  beta0 = 0
)
}
\description{
method for object of class "lm_robust"
}
