% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boottest}
\alias{boottest}
\title{S3 Method boottest}
\usage{
boottest(x, ...)
}
\arguments{
\item{x}{An object of type lm, fixest of felm}

\item{...}{other arguments}
}
\description{
S3 Method boottest
}
