% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boottest}
\alias{boottest}
\title{Fast wild cluster bootstrap inference}
\usage{
boottest(object, ...)
}
\arguments{
\item{object}{An object of type lm, fixest of felm}

\item{...}{other arguments}
}
\value{
An object of class \code{boottest}.
}
\description{
\code{boottest} is a S3 method that allows for fast wild cluster
bootstrap inference for objects of class lm, fixest and felm by  implementing
the fast wild bootstrap algorithm developed in Roodman et al., 2019.
}
\seealso{
\code{\link{boottest.lm}}, \code{\link{boottest.fixest}} and \code{\link{boottest.felm}}
}
