% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_r_lean}
\alias{check_r_lean}
\title{some checks when 'boot_algo = R-lean"}
\usage{
check_r_lean(weights, clustid, fe)
}
\arguments{
\item{weights}{NULL or numeric vector}

\item{clustid}{character vector}

\item{fe}{NULL or character scalar}
}
\description{
some checks when 'boot_algo = R-lean"
}
