% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data_1}
\alias{create_data_1}
\title{Function creates data for tests and examples}
\usage{
create_data_1(N, N_G, icc)
}
\arguments{
\item{N}{number of observations}

\item{N_G}{number of clusters}

\item{icc}{intra-cluster correlation}
}
\description{
Function creates data for tests and examples
}
