% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get_nthreads.R
\name{getBoottest_nthreads}
\alias{getBoottest_nthreads}
\title{get the number of threads for use with open mp}
\usage{
getBoottest_nthreads()
}
\value{
The number of threads currently used by boottest as set in options
}
\description{
get the number of threads for use with open mp
}
