% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_p_val_fwc.R
\name{invert_p_val.algo_oneclust}
\alias{invert_p_val.algo_oneclust}
\title{Inverts the bootstrap p-value and calculates confidence sets}
\usage{
\method{invert_p_val}{algo_oneclust}(
  object,
  point_estimate,
  se_guess,
  clustid,
  fixed_effect,
  X,
  Y,
  N,
  k,
  param,
  R0,
  B,
  v,
  Xr,
  alpha,
  beta0,
  W,
  n_fe,
  N_G
)
}
\arguments{
\item{object}{A regression object of class lm, feols or felm}

\item{point_estimate}{A scalar. Point estimate of the coefficient of interest from the regression model}

\item{se_guess}{A scalar vector of dimension 2. A guess of the standard error that initiates the p-value inversion.}

\item{clustid}{A vector with the clusters}

\item{fixed_effect}{Fixed effect to be projected out in bootstrap}

\item{X}{the design matrix with the (potentially demeand) covariates}

\item{Y}{A numeric vector containing the outcome variable}

\item{N}{An integer. Number of observations}

\item{k}{An integer. Number of covariates (excluding fixed effects that are projected out)}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{R0}{A vector with the test constraint. Dimension (numb_covariates - numb_fe) x 1. 0 for covariate "param", else 1}

\item{B}{An integer. Number of bootstrap iterations}

\item{v}{A matrix with draws from the bootstrap distribution}

\item{Xr}{A matrix, X*r}

\item{alpha}{A numeric between 0 and 1. Sets to confidence level: alpha = 0.05 returns 0.95\% confidence intervals}

\item{beta0}{A scalar. Shifts the null hypothesis.}

\item{W}{If the fixed_effect option is used,..., else NULL.}

\item{n_fe}{If the fixed_effect option is used, a scalar with the dimension of the fixed effect. Else NULL}

\item{N_G}{An integer. Number of clusters.}
}
\description{
Inverts the bootstrap p-value and calculates confidence sets
}
