% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{plot.boottest}
\alias{plot.boottest}
\title{Plot the bootstrap distribution of t-statistics}
\usage{
\method{plot}{boottest}(object)
}
\arguments{
\item{object}{An object of type boottest}
}
\description{
Plot the bootstrap distribution of t-statistics
}
