% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{preprocess2}
\alias{preprocess2}
\title{Fast wild cluster bootstrap inference}
\usage{
preprocess2(object, ...)
}
\arguments{
\item{object}{An objectect of type lm, fixest, felm or ivreg}

\item{...}{other arguments}
}
\value{
An object of class \code{preprocess2}.
}
\description{
\code{preprocess2} is a S3 method that fetches data from several model
objectects for use with \code{boottest()}.
}
