% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{preprocess2}
\alias{preprocess2}
\title{function that pre-processes regression objects of type lm, fixest and feols}
\usage{
preprocess2(object, cluster, fe, param, bootcluster, na_omit)
}
\arguments{
\item{object}{An object of class lm, fixest or felm}

\item{cluster}{A vector with the names of the clusters}

\item{fe}{A character scalar - fixed effect to be projected out, or NULL}

\item{param}{The univariate coefficients for which a hypothesis is to be tested}

\item{bootcluster}{The bootstrap sampling cluster.}

\item{na_omit}{Logical. If TRUE, \code{boottest()} omits rows with missing variables that are added to the model via the \code{cluster} argument in \code{boottest()}}
}
\value{
List containing preprocessed data for boottest estimation
}
\description{
function that pre-processes regression objects of type lm, fixest and feols
}
