% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setBoottest_engine}
\alias{setBoottest_engine}
\title{Sets the bootstrap algo to be run via \code{boottest()} and \code{waldboottest()}}
\usage{
setBoottest_engine(engine)
}
\arguments{
\item{engine}{Character scalar. Either 'R' or 'WildBootTests.jl'.
Default is 'R'}
}
\value{
No return value
}
\description{
Sets the bootstrap algo to be run via \code{boottest()} and \code{waldboottest()}
}
\examples{
\dontrun{
setBoottest_engine(engine = "R")
setBoottest_engine(engine = "WildBootTests.jl")
}
}
