% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{tidy.boottest}
\alias{tidy.boottest}
\title{summary <- function(x, ...){
#' S3 method to summarize objects of class boottest
#'@export
#'@param x object of type boottest
#'@param ... other arguments
UseMethod("summary", x)
}}
\usage{
\method{tidy}{boottest}(object, ...)
}
\arguments{
\item{object}{object of type boottest}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
plot <- function(x, ...){
#' S3 generic to plot bootstrap t statistics
#'@export
#'@param x object of type boottest
#'@param ... other arguments
UseMethod("plot", x)
}
S3 method to summarize objects of class boottest into tidy data.frame
}
