% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{tidy.boottest}
\alias{tidy.boottest}
\title{S3 method to summarize objects of class boottest into tidy data.frame}
\usage{
\method{tidy}{boottest}(object, ...)
}
\arguments{
\item{object}{object of type boottest}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A tidy data.frame with estimation results for objects of type
boottest
}
\description{
S3 method to summarize objects of class boottest into tidy data.frame
}
