% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_boottest.R
\name{tidy}
\alias{tidy}
\title{S3 method to summarize objects of class boottest into tidy data.frame}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{object of type boottest}

\item{...}{other arguments}
}
\description{
S3 method to summarize objects of class boottest into tidy data.frame
}
